update_tools/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
ca-certificates=2021.10.8=ha878542_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1d223b6_11
libgomp=11.2.0=h1d223b6_11
libzlib=1.2.11=h36c2ea0_1013
ncurses=6.2=h58526e2_4
openssl=3.0.0=h7f98852_2
pip=21.3.1=pyhd8ed1ab_0
python=3.9.7=hf930737_3_cpython
python_abi=3.9=2_cp39
readline=8.1=h46c0cb4_0
setuptools=59.2.0=py39hf3d152e_0
sqlite=3.36.0=h9cd32fc_2
telnetlib3=1.0.4=pypi_0
tk=8.6.11=h27826a3_1
tzdata=2021e=he74cb21_0
wheel=0.37.0=pyhd8ed1ab_1
xz=5.2.5=h516909a_1
zlib=1.2.11=h36c2ea0_1013
